week 12

+++ app/Http/Controllers/SupplierController.php
@@ -94,7 +94,7 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
-        // dd($supplier); 
+        $this->authorize('delete-permission');
         $supplier->delete();
         return redirect()->route('suppliers.index')->with('status', 'Data berhasil dihapus');
     }
